fix(extensions): review S3 downloads and split AWS S3 matchers
Treat s3api get-object as a local download, map put-object-annotation as
an object write, wrap the AWS catalog description, and split S3 matchers
out of the storage module so generated catalog fixtures stay in sync.
