+++ tests/guard_command_corpus_runner.py
@@ -18,7 +18,7 @@ from typing import TypedDict, cast
 
 EVALUATION_SHARD_COUNT = 4
 MAX_CONCURRENT_WORKERS = 4
-WORKER_TIMEOUT_SECONDS = 30
+WORKER_TIMEOUT_SECONDS = 60
 REPO_ROOT = Path(__file__).parents[1]
 SYNTHETIC_CWD = REPO_ROOT / "workspace"
 SYNTHETIC_HOME = REPO_ROOT / "home"
