+++ tests/test_guard_command_storage_extensions.py
@@ -172,6 +172,11 @@ STORAGE_REVIEW_CASES: tuple[tuple[str, str, str], ...] = (
         "command.storage.aws-s3.mb",
     ),
     (
+        "aws s3api get-object --bucket archive --key private.json outfile.json",
+        "AWS storage destructive command",
+        "command.storage.aws-s3.download",
+    ),
+    (
         "gcloud storage cp ./out gs://archive/private.json",
         "Google storage destructive command",
         "command.storage.google-cloud.cp",
@@ -209,7 +214,7 @@ STORAGE_SAFE_COMMANDS: tuple[str, ...] = (
     "aws s3 ls s3://archive",
     "aws --future-global-option account s3 ls s3://archive",
     "aws s3api list-objects-v2 --bucket archive",
-    "aws s3api get-object --bucket archive --key private.json outfile.json",
+    "aws s3api head-object --bucket archive --key private.json",
     "aws s3api head-bucket --bucket archive",
     "aws s3api put-object --generate-cli-skeleton input",
     "aws s3api delete-objects --generate-cli-skeleton output",
@@ -291,6 +296,7 @@ def test_storage_catalog_lists_supported_terminal_commands() -> None:
             "aws s3api put-bucket-versioning",
             "aws s3api put-object-tagging",
             "aws s3api get-object",
+            "aws s3api head-object",
         },
         "command.storage.google-cloud": {
             "gcloud storage rm",
