+++ tests/test_guard_extension_directory_operations.py
@@ -64,6 +64,7 @@ def test_aws_s3_operations_list_commands_and_default_floors() -> None:
         "command.storage.aws-s3.access-control",
         "command.storage.aws-s3.bucket-configuration",
         "command.storage.aws-s3.object-tagging",
+        "command.storage.aws-s3.download",
         "command.storage.aws-s3.get",
     }
     deletion = operations["command.storage.aws-s3.deletion"]
@@ -91,6 +92,8 @@ def test_aws_s3_operations_list_commands_and_default_floors() -> None:
     assert operations["command.storage.aws-s3.sync"]["commands"] == ["aws s3 sync"]
     assert "aws s3api create-bucket" in operations["command.storage.aws-s3.mb"]["commands"]
     assert "aws s3api put-object" in operations["command.storage.aws-s3.object-write"]["commands"]
+    assert operations["command.storage.aws-s3.download"]["commands"] == ["aws s3api get-object"]
+    assert operations["command.storage.aws-s3.download"]["defaultMode"] == "review"
     assert operations["command.storage.aws-s3.access-control"]["severity"] == "critical"
